+++ src/region/region.service.ts
@@ -3,13 +3,24 @@ import { CreateRegionDto, GetMongoIdDto } from './dto/create-region.dto';
 import { UpdateRegionDto } from './dto/update-region.dto';
 import { InjectModel } from '@nestjs/mongoose';
 import { Region } from './schema/region.schema';
-import { Model } from 'mongoose';
+import mongoose, { Model } from 'mongoose';
+import { User } from 'src/user/schema/user.schema';
 
 @Injectable()
 export class RegionService {
-  constructor(@InjectModel(Region.name) private RegionSchema: Model<Region>) {}
+  constructor(
+    @InjectModel(Region.name) private RegionSchema: Model<Region>,
+    @InjectModel(User.name) private UserSchema: Model<User>,
+  ) {}
 
-  async create(data: CreateRegionDto) {
+  async create(data: CreateRegionDto, req: any) {
+    let user = await this.UserSchema.findById(req.user.Id);
+    if (!user) {
+      return { Message: 'User malumotlari topilmadi' };
+    }
+    if (user.role != 'ADMIN') {
+      return { Message: "Regionni faqat admin qo'sha oladi" };
+    }
     let { name } = data;
     let region = await this.RegionSchema.findOne({ name });
     if (region) {
@@ -27,6 +38,9 @@ export class RegionService {
   }
 
   async findOne(id: string) {
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+      return { Message: "Region id Noto'g'ri" };
+    }
     let region = await this.RegionSchema.findById(id);
     if (!region) {
       return { Message: 'Not Found region' };
@@ -35,6 +49,9 @@ export class RegionService {
   }
 
   async update(id: string, data: UpdateRegionDto) {
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+      return { Message: "Region id Noto'g'ri" };
+    }
     let region = await this.RegionSchema.findById(id);
     if (!region) {
       return { Message: 'Not Found region' };
@@ -43,10 +60,13 @@ export class RegionService {
   }
 
   async remove(id: string) {
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+      return { Message: "Region id Noto'g'ri" };
+    }
     let region = await this.RegionSchema.findById(id);
     if (!region) {
       return { Message: 'Not Found region' };
     }
-    return this.RegionSchema.findByIdAndDelete(id)
+    return this.RegionSchema.findByIdAndDelete(id);
   }
 }
